modules/location-ingester-service/requirements.txt
kafka-python==2.0.2
grpcio-tools==1.62.0